split ModelBase
