+++ src/CapitalInvestments/PyomoModels/Factory.py
@@ -4,11 +4,6 @@
 Created on Feb. 21, 2019
 @author: wangc, mandd
 """
-#for future compatibility with Python 3-----------------------------------------
-from __future__ import division, print_function, unicode_literals, absolute_import
-import warnings
-warnings.simplefilter('default',DeprecationWarning)
-#End compatibility block for Python 3-------------------------------------------
 
 ################################################################################
 try:
